requirement.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.2
mysql-connector-python==8.0.33
google-cloud-vision==3.0.2
matplotlib==3.8.0
nltk==3.8.1
scikit-learn==1.3.1
